# features/v3.0.0-beta1/053-openapi-auth-and-identity/openapi-v2-key-auth-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BiSheng API-key API",
    "version": "2.0",
    "description": "HTTP and WebSocket endpoints authenticated with a Bearer API key."
  },
  "paths": {
...
    "/api/v2/assistant/chat/completions": {
      "post": {
        "tags": [
          "OpenAPI",
          "Assistant"
        ],
        "summary": "Assistant Chat Completions",
        "operationId": "assistant_chat_completions_api_v2_assistant_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIChatCompletionReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiBearer": []
          }
        ]
      }
    },
...
    "/api/v2/workstation/chat/completions": {
      "post": {
        "tags": [
          "OpenAPI",
          "WorkStation"
        ],
        "summary": "Daily Chat Completions",
        "operationId": "daily_chat_completions_api_v2_workstation_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDailyChatCompletionReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpenAIChatCompletionReq": {
        "properties": {
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Messages",
            "description": "Chat message list, only supporteduser、assistant。systemUse data from within the database"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "The only assistantID"
          },
          "n": {
            "type": "integer",
            "title": "N",
            "description": "Number of answers returned, The assistant side defaults to1, multiple answers are not supported at this time",
            "default": 1
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Whether to turn on streaming replies",
            "default": false
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "description": "Model Temperature, Incoming0or don't post means don't overwrite",
            "default": 0.0
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tools",
            "description": "Tools List, The assistant is temporarily unsupported, use the configuration of the assistant"
          }
        },
        "type": "object",
        "required": [
          "messages",
          "model"
        ],
        "title": "OpenAIChatCompletionReq"
      },
...
      "OpenDailyChatCompletionReq": {
        "properties": {
          "clientTimestamp": {
            "type": "string",
            "title": "Clienttimestamp"
          },
          "conversationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversationid"
          },
          "error": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "default": false
          },
          "generation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation",
            "default": ""
          },
          "isCreatedByUser": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iscreatedbyuser",
            "default": false
          },
          "isContinued": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iscontinued",
            "default": false
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "default": ""
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OpenDailyToolPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "search_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Enabled",
            "default": false
          },
          "parentMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentmessageid"
          },
          "overrideParentMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overrideparentmessageid"
          },
          "responseMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsemessageid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientTimestamp",
          "model"
        ],
        "title": "OpenDailyChatCompletionReq",
        "description": "Current APIChatCompletion minus task and knowledge-base selectors."
      },
      "OpenDailyToolPayload": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "default": 0
          },
          "tool_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Key"
          },
          "type": {
            "type": "string",
            "const": "tool",
            "title": "Type",
            "default": "tool"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OpenDailyToolPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OpenApiBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}